requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
pandas
openpyxl
pyarrow
python-dotenv
pytest
httpx